bugref:3809: Linux installer maintenance: do not manually delete device nodes in vboxadd.sh on stop unless the kernel modules were actually removed.  If the nodes are created by udev then they will not be recreated on next start, as the module will not be reloaded.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@65534 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/linux/installer/vboxadd.sh
@@ -270,10 +270,12 @@ stop()
         fail "Cannot unmount vboxsf folders"
     fi
     modprobe -q -r -a vboxvideo vboxsf vboxguest
-    egrep -q 'vboxguest|vboxsf|vboxvideo' /proc/modules &&
+    if egrep -q 'vboxguest|vboxsf|vboxvideo' /proc/modules; then
         info "You may need to restart your guest system to finish removing the guest drivers."
-    rm -f $userdev || fail "Cannot unlink $userdev"
-    rm -f $dev || fail "Cannot unlink $dev"
+    else
+        rm -f $userdev || fail "Cannot unlink $userdev"
+        rm -f $dev || fail "Cannot unlink $dev"
+    fi
     return 0
 }
